+++ src/main/java/com/shop/farmmunity/domain/item/dto/ItemFormDto.java
@@ -3,14 +3,23 @@ package com.shop.farmmunity.domain.item.dto;
 import com.shop.farmmunity.domain.item.constant.ItemClassifyStatus;
 import com.shop.farmmunity.domain.item.constant.ItemSellStatus;
 import com.shop.farmmunity.domain.item.entity.Item;
+import com.shop.farmmunity.domain.itemTag.entity.ItemTag;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.OneToMany;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
+import org.hibernate.annotations.LazyCollection;
+import org.hibernate.annotations.LazyCollectionOption;
 import org.modelmapper.ModelMapper;
 
 import java.util.ArrayList;
+import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 
 @Getter
@@ -35,9 +44,14 @@ public class ItemFormDto {
 
     private ItemClassifyStatus itemClassifyStatus;
 
-    @NotNull
     private Long postKeywordId;
 
+    @Builder.Default
+    @OneToMany(mappedBy = "item", cascade = CascadeType.ALL, orphanRemoval = true)
+    @LazyCollection(LazyCollectionOption.EXTRA)
+    Set<ItemTag> itemTags = new LinkedHashSet<>();
+
+    @NotBlank(message = "태그는 필수 입력 값입니다.")
     private String itemTagContents;
 
     private List<ItemImgDto> itemImgDtoList = new ArrayList<>(); //수정 시, 상품 이미지 정보 저장하는 리스트
@@ -54,4 +68,19 @@ public class ItemFormDto {
         return modelMapper.map(item, ItemFormDto.class);
     }
 
+    public String getExtra_itemTagLinks() {
+        return itemTags
+                .stream()
+                .map(itemTag -> {
+                    String text = "#" + itemTag.getItemKeyword().getContent();
+
+                    return """
+                            <a href="%s" class="text-link">%s</a>
+                            """
+                            .stripIndent()
+                            .formatted(itemTag.getItemKeyword().getListUrl(), text);
+                })
+                .sorted()
+                .collect(Collectors.joining(" "));
+    }
 }
