+++ src/main/resources/templates/item/itemDtl.html
@@ -222,11 +222,10 @@
                 </div>
             </div>
 
-
         </div>
     </div>
 
-    <div th:utext="${itemEntity.extra_itemTagLinks}"></div>
+    <div th:utext="${item.extra_itemTagLinks}"></div>
 
     <div class="jumbotron jumbotron-fluid mgt-30">
         <div class="container">
